services/email_watcher/requirements.txt
flask
requests
google-api-python-client
google-auth
google-cloud-secret-manager
google-cloud-storage
google-cloud-firestore